+++ src/karaoke/karaoke.module.ts
@@ -4,7 +4,7 @@ import { HttpModule } from '@nestjs/axios';
 import { SearchSongsUC } from './application/use-cases/search-songs.uc';
 import { SpotifyRepository } from './infrastructure/spotify/spotify.repository';
 import { RedisCache } from './infrastructure/cache/redis.cache';
-import { SpotifyService } from './infrastructure/spotify/spotify.service';
+import { SpotifyAuthService } from './infrastructure/spotify/spotify-auth.service';
 import { AuthModule } from 'src/auth/auth.module';
 
 @Module({
@@ -13,7 +13,7 @@ import { AuthModule } from 'src/auth/auth.module';
   providers: [
     SearchSongsUC,
     { provide: 'SongRepository', useClass: SpotifyRepository },
-    SpotifyService,
+    SpotifyAuthService,
     RedisCache,
   ],
 })