+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - Add architecture DDD with clean architecture in domain Karaoke.
 - Add service Redis to save token generate from Spotify, then request song to Spotify API.
+- Add unit test of Karaoke module.